update component;add Form;
