+++ src/components/Layout/Layout.tsx
@@ -1,16 +1,8 @@
-import {
-  Layout as Container,
-  InputSearch,
-  Sidebar,
-
-} from '@ebs-integrator/react-ebs-ui';
-
-
-
+import { Layout as Container, InputSearch, Sidebar } from '@ebs-integrator/react-ebs-ui';
+import { NavLink } from 'react-router-dom';
 const Layout: React.FC = ({ children }) => {
   return (
-
-    <Container >
+    <Container>
       <Container.Topbar>
         <Container.Topbar.Toggler />
 
@@ -24,26 +16,27 @@ const Layout: React.FC = ({ children }) => {
       </Container.Topbar>
 
       <Sidebar>
-        <Sidebar.TopMenu>
+        <Sidebar.TopMenu></Sidebar.TopMenu>
 
-        </Sidebar.TopMenu>
-        Sidebar text
+        <Sidebar.Options>
+          <Sidebar.Options.TopSide>
+            <NavLink to="/">
+              <Sidebar.Options.Item text="initial" />
+            </NavLink>
+            <NavLink to="/dashboard">
+              <Sidebar.Options.Item text="advance" />
+            </NavLink>
+            <NavLink to="/form">
+              <Sidebar.Options.Item text="form" />
+            </NavLink>
+          </Sidebar.Options.TopSide>
+        </Sidebar.Options>
       </Sidebar>
 
-      <Container.Content>
-        {children}
-      </Container.Content>
+      <Container.Content>{children}</Container.Content>
 
       <Container.Footer />
-
     </Container>
-
-
   );
 };
 export default Layout;
-
-
-
-
-
