+++ src/components/index.scss
@@ -2,15 +2,9 @@
 @import './Overlay/Overlay.scss';
 @import './Form/Form.scss';
 @import './Layout/Layout.scss';
-@import './Dashboard/Dashboard.scss';
-@import './Congrats/Congrats.scss';
-@import './CartPrize/CartPrize.scss';
 @import './Cart/Cart.scss';
 @import './Recharts/Recharts.scss';
-@import './CartEmployee/CartEmployee.scss';
 @import './AvatarListPic/AvatarListPic.scss';
-@import './CartMeetUp/CartMeetUp.scss';
-@import './CartProfile//CartProfile.scss';
 @import './Timeline//Timeline.scss';
 @import './ProfileMedia/ProfileMedia.scss';
 @import './ImageMedia/ImageMedia.scss';
