+++ src/features/dashboard/CartPrize/CartPrize.scss
