+++ src/features/dashboard/index.scss
@@ -1,0 +1,5 @@
+@import './Congrats/Congrats.scss';
+@import './CartEmployee/CartEmployee.scss';
+@import './CartPrize/CartPrize.scss';
+@import './CartMeetUp/CartMeetUp.scss';
+@import './CartProfile//CartProfile.scss';
