--- src/features/home/ChatView/ChatView.tsx
@@ -1,8 +1,0 @@
-const ChatView = () => {
-  return (
-    <div>
-      <p>ChatView</p>
-    </div>
-  );
-};
-export default ChatView;
