+++ src/features/home/FormRepeater/FormRepeater.scss
@@ -1,0 +1,55 @@
+.box-form {
+  padding: rem(20px);
+  margin-bottom: rem(28px);
+  border-radius: rem(5px);
+  box-shadow: 0px 0px 25px -5px $primary-200;
+
+  background-color: $white;
+}
+
+.form-repeater {
+  margin: 0;
+  padding: 0;
+}
+
+.form-line {
+  width: 100%;
+  height: 1px;
+  margin-top: rem(14px);
+  margin-bottom: rem(14px);
+  background-color: $grey-200;
+}
+
+.ebs-button--ghost.ebs-button__wrapper.delete-button {
+  height: rem(40px);
+  margin-top: rem(21px);
+  color: $danger-color;
+  border-color: $danger-color;
+}
+
+.ebs-button--ghost.ebs-button__wrapper.delete-button:hover,
+.ebs-button--ghost.ebs-button__wrapper.delete-button:focus {
+  background-color: $grey-200;
+}
+
+.ebs-button--ghost.ebs-button__wrapper.delete-button .delete-text-button {
+  margin-top: rem(2px);
+}
+
+.ebs-button--ghost.ebs-button__wrapper.delete-button .delete-text-button,
+.ebs-button--ghost.ebs-button__wrapper.delete-button .delete-button {
+  color: $danger-color;
+}
+
+// .ebs-button--ghost.ebs-button__wrapper.delete-button:hover .delete-text-button,
+// .ebs-button--ghost.ebs-button__wrapper.delete-button:focus .delete-text-button {
+//   color: $danger-color;
+// }
+
+.box-text-button {
+  display: flex;
+  align-items: center;
+}
+.text-button {
+  margin-left: rem(3px);
+}
