--- src/features/home/HomeView/HomeView.tsx
@@ -1,9 +1,0 @@
-const HomeView = () => {
-  return (
-    <div>
-      <p>HomeView</p>
-    </div>
-  );
-};
-
-export default HomeView;
