+++ src/resources/index.ts
@@ -1,11 +1,12 @@
-export { ReactComponent as ebsImage } from "./ebs-image.svg";
-export { ReactComponent as question } from "./question.svg";
-export { ReactComponent as facebook } from "./facebook.svg";
-export { ReactComponent as github } from "./github.svg";
-export { ReactComponent as mail } from "./mail.svg";
-export { ReactComponent as twitter } from "./twitter.svg";
-export { ReactComponent as flagLock } from "./flag-lock.svg";
-export { ReactComponent as medal } from "./medal.svg";
-export { ReactComponent as geolocation } from "./geolocation.svg";
-export { ReactComponent as calendar } from "./calendar.svg";
-
+export { ReactComponent as ebsImage } from './ebs-image.svg';
+export { ReactComponent as question } from './question.svg';
+export { ReactComponent as facebook } from './facebook.svg';
+export { ReactComponent as github } from './github.svg';
+export { ReactComponent as mail } from './mail.svg';
+export { ReactComponent as twitter } from './twitter.svg';
+export { ReactComponent as flagLock } from './flag-lock.svg';
+export { ReactComponent as medal } from './medal.svg';
+export { ReactComponent as geolocation } from './geolocation.svg';
+export { ReactComponent as calendar } from './calendar.svg';
+export { ReactComponent as plusAdd } from './plus-add.svg';
+export { ReactComponent as xDelete } from './x-delete.svg';
