+++ src/styles/index.scss
@@ -5,4 +5,7 @@
 
 @import '../features/auth/LoginPage/LoginPage.scss';
 
+@import '../features/dashboard/index.scss';
+
+@import '../features/home/FormRepeater/FormRepeater.scss';
 @import '../components/index.scss';
